Update docstring position in extraction_to_txt_file methods
The commit relocates the docstrings to be directly under the method definition, as per Python's standard practice, in the 'extraction_to_txt_file' script. These provide information about the functions 'get_data' and 'save_quality_docs'. This revision ensures that the method's purpose is clarified to the developers in the correct position as a part of Python's best practices. This doesn't affect functionality but improves script readability.
Removed .gitignore file from repo